requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
joblib==1.0.1
nltk==3.5
oauthlib==3.1.0
PySocks==1.7.1
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
textblob==0.15.3
tqdm==4.59.0
tweepy==3.10.0
urllib3==1.26.3
